chore(*): messages benchmark cleanup
Cleanup on messages benchmark helper.
